+++ src/component/index.js
