+++ src/component/utils/PrivateStore.js
