+++ src/l/index.js
@@ -1,8 +1,7 @@
 import Listener from './Listener';
 
 const l = (component, key, childPath = '') => {
-  const listener = new Listener(component, key, childPath);
-  return listener;
+  return new Listener(component, key, childPath);
 };
 
 export default l;
