+++ src/r/utils/isNode.js
@@ -1,0 +1,2 @@
+const isNode = a => !!(a && a.nodeType);
+export default isNode;
