+++ src/r/utils/listenerToNode.js
@@ -1,0 +1,8 @@
+import isNode from './isNode';
+
+const listenerToNode = (value) => {
+  if (isNode(value)) return value;
+  return document.createTextNode(value);
+};
+
+export default listenerToNode;
