+++ src/register.js
