+++ src/utils/__tests__/ComponentClass.test.js
@@ -1,5 +1,5 @@
 import Component from '../ComponentClass';
-import r from '../r';
+import r from '../../r';
 
 /** @jsx r **/
 describe('ComponentClass.js', () => {
